feat: improve settings page
