--- .hintrc
@@ -1,6 +1,0 @@
-{
-  "extends": ["development"],
-  "hints": {
-    "no-inline-styles": "off"
-  }
-}
