+++ apps/cms/src/components/layout/wrapper.tsx
@@ -11,7 +11,7 @@ export function WorkspacePageWrapper({
   return (
     <div
       className={cn(
-        "flex flex-col h-full w-full max-w-5xl mx-auto py-8",
+        "flex flex-col h-full w-full max-w-4xl mx-auto py-8",
         className,
       )}
     >
