+++ apps/cms/src/lib/actions/user.ts
@@ -1,0 +1,1 @@
+import "server-only";
